preparing documentation
