+++ huginn-net-http/src/http.rs
@@ -1,13 +1,3 @@
-//! Pure HTTP fingerprint data types.
-//!
-//! These types are intentionally free of any matching/database logic so the
-//! crate can stay independent of `huginn-net-db`. The database crate
-//! re-exports them.
-
-/// Version of the HTTP protocol used in a request or response.
-///
-/// Used in signatures to distinguish behavior between HTTP/1.0 and HTTP/1.1.
-/// The `Any` variant is used in database signatures to match any HTTP version.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub enum Version {
     /// HTTP/1.0
